requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-rest-knox==4.2.0
django-q==1.3.9
celery==5.3.1
gunicorn==20.1.0
